backend/requirements.txt
# SME-Pulse AI Backend Dependencies
# FastAPI + PostgreSQL + AI Integration

fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pandas==2.1.3
numpy==1.26.2
pydantic==2.5.2
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==41.0.7
python-docx==1.1.0
openpyxl==3.1.2
pdfplumber==0.10.3
httpx==0.25.2
aiofiles==23.2.1
python-multipart==0.0.6

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Development
black==23.11.0
isort==5.12.0
mypy==1.7.1
flake8==6.1.0

